requirements.txt
anthropic>=0.40.0
crawl4ai>=0.4.0
outscraper>=4.0.0
supabase>=2.0.0
pydantic>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
asyncio-throttle>=1.0.0
